+++ pcc_colonel_blotto/__init__.py
@@ -1,3 +1,3 @@
 """PCC Colonel Blotto."""
 
-__version__ = "0.5.0"
+__version__ = "0.6.0"